+++ src/app/features/user-roles/page/create-page/create-page.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { FormComponent } from '@features/user-roles/component/form/form.component';
+
+@Component({
+  selector: 'app-create-page',
+  imports: [FormComponent],
+  templateUrl: './create-page.component.html',
+  styleUrl: './create-page.component.scss'
+})
+export class CreatePageComponent {
+  submit() {}
+}