+++ src/app/features/user-roles/page/update-page/update-page.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { FormComponent } from '@features/user-roles/component/form/form.component';
+
+@Component({
+  selector: 'app-update-page',
+  imports: [FormComponent],
+  templateUrl: './update-page.component.html',
+  styleUrl: './update-page.component.scss'
+})
+export class UpdatePageComponent {
+  submit() {}
+}