+++ src/app/features/user-roles/user-role.routes.ts
@@ -1,11 +1,11 @@
 import { Routes } from '@angular/router';
 import { CrudRoutes } from '@shared/constant/navigation-route.const';
 
-export const manageUserRoutes: Routes = [
+export const userRolesRoutes: Routes = [
   {
     path: '',
     loadComponent: () =>
-      import('@features/manage-user/page/list-page/list-page.component').then(
+      import('@features/user-roles/page/list-page/list-page.component').then(
         c => c.ListPageComponent
       )
   },
@@ -13,13 +13,13 @@ export const manageUserRoutes: Routes = [
     path: CrudRoutes.CREATE,
     loadComponent: () =>
       import(
-        '@features/manage-user/page/create-page/create-page.component'
+        '@features/user-roles/page/create-page/create-page.component'
       ).then(c => c.CreatePageComponent)
   },
   {
     path: CrudRoutes.DETAIL,
     loadComponent: () =>
-      import('@features/manage-user/page/view-page/view-page.component').then(
+      import('@features/user-roles/page/view-page/view-page.component').then(
         c => c.ViewPageComponent
       )
   },
@@ -27,7 +27,7 @@ export const manageUserRoutes: Routes = [
     path: CrudRoutes.EDIT,
     loadComponent: () =>
       import(
-        '@features/manage-user/page/update-page/update-page.component'
+        '@features/user-roles/page/update-page/update-page.component'
       ).then(c => c.UpdatePageComponent)
   }
 ];