+++ src/app/shared/component/table/index.ts
@@ -1,0 +1,4 @@
+import { TableColumnDirective } from './directives/table-column.directive';
+import { TableComponent } from './table.component';
+
+export { TableComponent, TableColumnDirective };