+++ src/app/shared/constant/error-message.const.ts
@@ -30,5 +30,33 @@ export const ErrorMessageConst = {
       message: 'Please enter password'
     },
     { type: 'mismatch', message: 'Passwords do not match' }
+  ],
+  NAME: [
+    {
+      type: 'required',
+      message: 'Please enter name'
+    },
+    {
+      type: 'minlength',
+      message: 'Minimum {0} characters required'
+    },
+    {
+      type: 'maxlength',
+      message: 'Maximum {0} characters allowed'
+    }
+  ],
+  LABEL: [
+    {
+      type: 'required',
+      message: 'Please enter label'
+    },
+    {
+      type: 'minlength',
+      message: 'Minimum {0} characters required'
+    },
+    {
+      type: 'maxlength',
+      message: 'Maximum {0} characters allowed'
+    }
   ]
 };