+++ src/app/shared/utils/action-utils.ts
@@ -1,0 +1,18 @@
+export class ActionUtils {
+  static readonly viewBtnConfig = {
+    image: {
+      src: 'assets/view-icon.svg'
+    }
+  };
+  static readonly updateBtnConfig = {
+    image: {
+      src: 'assets/update-icon.svg'
+    }
+  };
+
+  static readonly deleteBtnConfig = {
+    image: {
+      src: 'assets/delete-icon.svg'
+    }
+  };
+}